Move brush chart only to AnalysisFilterChart
